refactor: update types for shared ui components and divide apiSlice
