+++ src/shared/types/chooseTypes.ts
@@ -1,6 +1,6 @@
-export interface IChooseProps {
-  id: string;
+import { InputHTMLAttributes } from 'react';
+
+export interface IChooseProps extends InputHTMLAttributes<HTMLInputElement> {
   isChecked: boolean;
   onChoose: () => void;
-  className?: string;
 }
